ARM: msm: hammerhead: Ignore watchdog init if device is interactive
Workaround a race condition in wifi driver where we init
the watchdog while we are receiving packets and vote for
resume but the wakelock is not long enough to get the fw
up and running. The device falls asleep and at some point
we fail to properly complete the waitqueue and trigger a
wdog bite.

Instead of messing around with longer packet timeouts,
just ignore watchdog inits while the device is already
interactive.

Change-Id: I7d1b31c9aafef0a433ad6f510e0c8e905265cdc6

Signed-off-by: franciscofranco <franciscofranco.1990@gmail.com>